Add symbol table
The idea was to reduce memory usage, however the savings are not that
great: for creatives.bert, the memory usage drop from 490M to 474M,
not nearly as much as I thought could be gained.

I will not merge this branch, but I'll leave it in case I want to
revisit in the future.