Add new bert term, SmallAtom
The goal of SmallAtom is to use less memory for atoms of 8 characters
or less.  On creatives.bert, my largest bert file (96M), the maximum
memory usage of ppbert drops from 497M to 488M.  It's something, but I
don't think it's enough to warrant an extra case, so I'll leave this
banch dormant and re-open it in the future if necessary.