Use a simpler method to compute indentation
The goal of this branch was to see if a change to the way indent
strings are computed would affect performance.  On the largest bert
file (creatives.bert, 96M) in my test set, the performance drops from
~3.8s to ~3.74s.  The change is too minute to merge, especially
considering that there is no more indentation past a certain level.
I'll keep this branch alive in case we need to revisit in the future.